Use a different service for vmcore dump.
During the dump itself, spurious CPU or memory hotplug events will cause
the dump to fail, because the service would be restarted, and an old
incomplete dump would prevent the dump from being collected.

This work also allows us to stop requiring network during kdump loading,
and sets the way to stop requiring network when the dump is not over the
network.

Signed-off-by: Thadeu Lima de Souza Cascardo <cascardo@debian.org>