Allow proper reload of kdump after multiple hotplug events.
When a CPU is hotplugged, multiple events will be issued for each CPU
thread getting online. On a POWER system, that usually means 8 threads.
Those 8 events will cause systemd to consider the multiple restarts as
failed.

One alternative fix would be setting StartLimitIntervalSec to 0, but
that would apply to all cases where those failures might happen, not
only on the hotplug case.

Instead, we use reset-failed before try-restart, which will allow those
multiple restarts to happen on a short interval.

Signed-off-by: Thadeu Lima de Souza Cascardo <cascardo@debian.org>