Use kdump-config reload after cpu or memory hotplug.
The solution used to allow multiple reloads during a hotplug event ended
up not working because udev won't execute two commands under a shell.

So, using a single command that reloads and do not interact with systemd
should work here.

As an extra, this adds support for other init systems.

Signed-off-by: Thadeu Lima de Souza Cascardo <cascardo@debian.org>