kdump-tools: Cope with the lack of initramfs-tools OPTION=VAR feature in Debian
Currently, there's an interesting feature on Ubuntu's initramfs-tools
that allows a fine-tuned control on the hooks/scripts that will be
executed/included in the initrd image, depending on options set on
initramfs-tools.conf and validated via OPTION=VAR in the hooks/scripts.

Since this isn't included in Debian initramfs-tools (for some unknown
reason), we need to amend the sysctl handling on Debian's kdump-tools
by adding more checks in both the hook and script, in order to avoid
them to run in regular boot.

The idea of having an extra patch for that is (a) to keep better code
sync with Ubuntu and commit organization, (b) expectation on including
the OPTION=VAR feature in Debian's initramfs-tools, hence this patch
might be reverted in the future.

Required-by-commit: "kdump-tools: Add sysctl override mechanism"
Signed-off-by: Guilherme G. Piccoli <gpiccoli@canonical.com>