kdump_save_core_{ftp,ssh}: Use unique temp dmesg files
Both functions attempt to use the same temp file for staging dmesg output
for transfer. Now that we've removed the limitation on dumping usning only
1 protocol at a time, it's possible that the ftp function has already
written the file when the ssh function runs. This will fail because
makedumpfile will refuse to overwrite an existing file. Avoid this
by using unique file paths.

Fixes: e8fb28e72 ("Support dumping kernel via both SSH and NFS")
Signed-off-by: dann frazier <dannf@debian.org>