Remove path from update-grub in maintainer scripts
The indicated program run in a maintainer script has a prepended path.
Programs called from maintainer scripts normally should not have a
path prepended. dpkg ensures that the PATH is set to a reasonable
value, and prepending a path may prevent the local administrator from
using a replacement version of a command for some local reason.

If the path is used to test a program for existence, please use if
which $program > /dev/null; then ….

If you intend to override this tag, please make sure that you are in
control of the installation path of the according program and that you
won't forget to change this maintainer script, too, if you ever move
that program around.

Refer to Debian Policy Manual section 6.1 (Introduction to package
maintainer scripts), Debian Developer's Reference section 6.4 (Best
practices for maintainer scripts), Bug#769845, Bug#807695, and
https://lists.debian.org/debian-devel/2014/11/msg00044.html for
details.

Signed-off-by: Benjamin Drung <bdrung@debian.org>
Signed-off-by: dann frazier <dannf@debian.org>