kdump-tools.service: Start kdump service a bit earlier in boot
The kdump systemd service currently is started in the multi-user
systemd target - this behavior was introduced by commit
051fa5dd ("Use multi-user.target.").

There is no big problem in that, but for the purpose of the
kdump memory estimator in the subsequent patches, it's better
for memory estimation accuracy to start the service a bit early
in the boot time (and also there is the bonus of enabling the
panic kexec a bit before, covering a slightly bigger window).

This shouldn't have any noticeable effect for the users.
Notice that we guarantee local-fs and networking services are
enabled before kdump.

Signed-off-by: Guilherme G. Piccoli <gpiccoli@igalia.com>