Disable kernel hook initramfs generation when running in a chroot
By default, we generate the kdump initramfs using MODULES=dep to
save space. But initramfs-tools is unable to calculate the list
of required modules in this environment:

  /etc/kernel/postinst.d/kdump-tools:
  kdump-tools: Generating /var/lib/kdump/initrd.img-5.15.0-1040-nvidia
  mkinitramfs: failed to determine device for /
  mkinitramfs: workaround is MODULES=most, check:
  grep -r MODULES /var/lib/kdump/initramfs-tools

  Error please report bug on initramfs-tools
  Include the output of 'mount' and 'cat /proc/mounts'
  update-initramfs: failed for /var/lib/kdump/initrd.img-5.15.0-1040-nvidia with 1.
  run-parts: /etc/kernel/postinst.d/kdump-tools exited with return code 1

And even if it could, it is likely that the chroot environment is running
on some build system and not the intended target, and thus the wrong
set of modules would be included. Just skip it for now. The systemd service
will detect the missing initrd at boot and generate one then.

LP: #2043059