initrd: Keep track of the kdump minimal initrd size during its creation
The goal for this specific change is to keep track of the minimal kdump
initrd size during the creation time, to be used in kdump memory estimation
heuristics. Doing that on initramfs create time takes a bit less effort;
otherwise we would need to decompress the minimal initrd to account for
its exact size or make use of (variable) compression tools to gather its
decompressed size.
It's much easier to just calculate that upfront with the aid of mkinitramfs,
which allow us to keep the temporary rootfs of the initrd, so we just collect
the size in a /var/lib/kdump file and erase the temporary rootfs after.

Signed-off-by: Guilherme G. Piccoli <gpiccoli@igalia.com>