kdump-tools: Introduce a kdump crashkernel memory estimator
This patch introduces a memory estimator for kdump-tools. It's a known
(long-term) challenge for kdump utilities to estimate / anticipate the
amount of needed memory for a complete / correct kernel crash dump.
From one side, we don't want to waste system usable memory and set a
crashkernel reservation too big, which would allow a correct kdump but
at the same time, subtracts a big amount of memory from the users.
On the other hand, setting a small memory reservation on crashkernel
parameter might lead to OOM condition in kdump, potentially wasting the
(only?) opportunity to collect a valid crash dump, which could be the
last resource to debug a hard-to-reproduce issue.

Some attempts were made in the past to automate the setting for the
crashkernel parameter - [0][1] are known ones, rejected from community but
merged in certain distros. Relying in the kernel entirely seems to
be far from ideal (like [0] and [1]), this is naturally a multi-layer
issue. The policy (or suggestion) for the proper amount to reserve in
crashkernel should be done in userspace, but kernel is required in order
to collect information about memory usage.

This patch introduces an approach based in a simple heuristic: we need
to have enough free memory in kdump environment to boot the system
(which requires far less memory than a regular boot, due to limitations
like nr_cpus=1, etc) and, after the boot, collecting the dump using an
userspace tool (like makedumpfile). For that, we need to account the
following components:

(a) kernel image (vmlinux), and its decompressed portion (if a
compressed image is used);

(b) initrd image + its decompressed size (mostly freed after usage,
if the kdump process happens at the rootfs level like it currently
does in Debian/Ubuntu);

(c) the memory used for the system boot itself (basic drivers, udev
discovery, systemd, etc);

(d) memory used by the userspace tool (makedumpfile in our case).

(e) surplus memory that is consumed in a regular boot but not in
kdump boot (specially due to nr_cpus=1 and restricted memory, which
implies less memory spent in CPU structs, page tables and generic
memory management structs);

Below we'll describe how the implementation calculates (or empirically
estimates) all the pieces. The basic heuristic formula is:

crashkernel = (a) + (b) + (c) + (d) - (e)

For (a), we rely in 2 parts: first we take account of the vmlinux size,
and then we multiply by an estimation of the uncompressed size factor;
this was collected in an experiment (see [2]). We consider only the
vmlinux decompressed size for the heuristics, since we assume the
decompressed portion replaces the compressed one in the RAM.

For (b), we added code in a previous patch to preserve the temporary
files in the kdump initrd generation and collect its decompressed size
in a file; so we account the total initrd size by adding the compressed
initrd size with the decompressed previous-calculated size. Notice that
this portion will be mostly freed after usage, so we also account it as
(d) - the free memory needed by makedumpfile, for kdump collection.
Empirically we noticed that adding more memory as (d) would be an
overestimation.

Regarding (c), this implementation makes use of a boot time script to
collect system used memory in a regular boot - by saving /proc/meminfo
in a file, we can later calculate the used memory in that time by
following a calculation based on the procfs "free" tool estimation. We
also take hugepages into account, since they are not used in kdump but
accounted as used memory by the system in a regular boot. The formula
for used memory is:

UsedMem = MemTotal - MemFree - Buffer - Cached - SReclaimable - Hugetlb

Important to notice that we could have used a kdump boot to collect
/proc/meminfo, making the calculation very precise, but then the
estimator would be useless unless a first kdump is successful, and also
the configuration might change, so the option was to account used memory
in every *regular boot*, which has the downside of introducing portion
(e) to the equation - the "rounding" error, i.e., memory that is used
in regular boot but not in kdump boot.

To calculate precisely (e) would be practically impossible, so we rely in
some experiments. Basically, we noticed that the more CPUs, Memory (RAM)
and Devices (PCIe mainly) the system has, more memory is consumed. We
usually do not disable devices in a kdump environment, so the PCIe devices
part is out of our estimation. The details on how memory increases as we
add CPUs/RAM are in [3]. In the end, we used the empirically calculated
values from [3] in order to adjust the crashkernel estimated value
based on the CPU number (which is restricted to 1 in kdump boot) and
based on the RAM memory of the machine. Currently, this adjustment is
only performed for x86 architecture (32/64 bit), but we plan to improve
the tests for other architectures in the future.

Finally, in case the estimator understimates the memory required due to
flaws in the heuristic portion (e), we have a fallback: we consider that
the vmlinux and initrd (both decompressed) plus some extra fixed amount
of memory would be the bare minimum required for kdump, hence the estimator
never suggests less memory than this minimum.

With all that said, no functional change is done specifically by this
patch, other than doing some operations on "kdump-config show" in order
to show the estimation value. We consider the "show" operation as the
best place in which the estimation output fits.

[0] https://lore.kernel.org/lkml/20090904101047.5220.20618.sendpatchset@localhost.localdomain
[1] https://lore.kernel.org/linux-mm/20210507010432.IN24PudKT%25akpm@linux-foundation.org

[2] The vmlinux experiment: using binwalk or the scripts/extract-vmlinux in the
kernel tree, we investigated the vmlinux compressed size in Debian and Ubuntu,
for multiple architectures.

*** Debian ***
AMD64:
* vmlinuz-4.19.0-6-amd64 (xz)
compressed: 5270768
uncompressed: 27747956
ratio: 5.265

* vmlinuz-6.5.0-5-amd64 (xz)
compressed: 9127904
uncompressed: 68040660
ratio: 7.454

*** Ubuntu ***
AMD64:
* vmlinuz-5.4.0-33-generic (lz4)
compressed: 11660160
uncompressed: 45035736
ratio: 3.863

* vmlinuz-5.4.0-33-lowlatency (lz4)
compressed: 11709312
uncompressed: 45018936
ratio: 3.845

* vmlinuz-6.5.0-14-generic (zstd)
compressed: 14264576
uncompressed: 75013396
ratio: 5.259

ARM64:
* vmlinuz-4.15.0-101-generic (gzip)
compressed: 8448185
uncompressed: 23759360
ratio: 2.813

* vmlinuz-5.4.0-33-generic (gzip)
compressed: 10536506
uncompressed: 30116352
ratio: 2.859

S390X:
* vmlinuz-4.15.0-101-generic (xz)
compressed: 4440064
uncompressed: 15456256
ratio: 3.481

* vmlinuz-5.4.0-33-generic (lz4)
compressed: 8089600
uncompressed: 17090216
ratio: 2.113

PPC64:
-> Not compressed *apparently*
(see https://github.com/ReFirmLabs/binwalk/issues/467).

[3] The CPU/Memory experiment: This was a simple experiment in which
we added more CPUs to a VM and collected the used memory in boot time;
after that, deltas were calculated to determine empirically the
nonlinear decreasing increment in memory usage the more CPUS we
added to the machine. A discrete "modeling" was then implemented
based on such data. Similar experiment was performed for memory;
the more RAM we add to a VM, more memory is consumed, but it's not
linear.

The experiment is full of values / deltas, so wouldn't be ideally
translated to a commit message. It's available in a spreadsheet
(OpenDocument Spreadsheet format) here:
https://guilhermepiccoli.github.io/kdump-estimator-cpu_mem-experiment_2020-11-05.ods

Signed-off-by: Guilherme G. Piccoli <gpiccoli@igalia.com>