Delete the temporary sysctl files on kernel removal
Commit 4501c234 ("kdump-tools: Add sysctl override mechanism") introduced
a sysctl handling mechanism for kdump-tools. This mechanism relies on a
file created on /var/lib/kdump, to track sysctls we should enforce when
booting a kdump kernel.

The problem is that such commit forgot to add a removal hook for the
sysctl transient file on kernel removal, so we ended up never deleting
these files (that would sum one per installed/booted kernel) - let's fix
it by adding such removal hook.

Fixes: 4501c234 ("kdump-tools: Add sysctl override mechanism")
Signed-off-by: Guilherme G. Piccoli <gpiccoli@igalia.com>